requirements.txt
python-twitch-client==0.7.1
twitchio==1.0.0
xmltodict==0.12.0
